console: Move landing view logic to app view
