+++ pkg/webui/console/views/app/app.styl
@@ -20,6 +20,7 @@
 .main
   display: flex
   flex: 1 1 auto
+  position: relative
   +media-query($bp.s)
     flex-direction: column
 
@@ -27,6 +28,9 @@
   overflow: auto
   flex: 1
   position: relative
+
+.stage
+  padding-bottom: $ls.xxs
 
 .mobile-breadcrumbs
   display: none
